Created CAT08 utilities and script to centerlines -> graph
The idea is to find the indices, for each centerline, at which it departs from another centerline.
Get the higher index for each couple of centerlines, and assume that
the couple is connected (near enough) for each index.
Then, try a rudimental mean shift which takes into account the split region.
It works but yields rudimental results, the split is not beautiful and smooth.