Progresses on creating a graph from cat08 centelrines
Here, I divided manually the centerlines into segments,
each of which has its cathegory label.
The segments common to more centerlines have
their unique labels, and so do standalone segments.
Then, in each common segment all the centerlines
contributing to that segment are selected to apply
a mean to them, sort of like mean shift but much more basaic.
Result is satisfactory.
Now I have N disjointed segments, each of which
is a list of points ordered from the one higher up in the
path (nearest to ostium) to the most distal.

Now, the idea will be to smooth out the mean-ed segments,
to connect the adjacent segments (which can be done
either with prior connection knowledge or automatically quite easily)
and to make a graph out of all of this.

Manual annotation performed just for image 00 and 01,
the rest is still to be done.